[css-text] Avoid mid-word breaking if there are previous opportunities
Leading white-spaces are indeed breaking opportunities that should
prevent, if there are no other css properties forcing it, breaking text
in the middle of a word.

We are doing so if the leading white-space sequence is longer than 1
character, but when we have a single leading white-space, we are missing
that breaking opportunity and we may lead to cases, like the one
described in the bug, where the word is broken to avoid an overflow,
honoring the overflow-wrap property.

This change tries to solve the issue by preventing to execute the last
part of the BreakingContext::HandleText logic, where we try to rewind
to the previous mid-word breaking opportunity. The idea is to detect
white-space breaking opportunities in the PrepareForNextCharacter
function that indicates that it doesn't make sense to break the word.

Bug: 854624
Change-Id: I49ec8282e899a32990662c104b48cab04057b909