Remove webkit-appearance.tentative.html WPT test
This test is not a good indicator of -webkit-appearance support; it only
checks whether computed-value == applied-value for -webkit-appearance
values specified on a vanilla <div>. In reality most values of
-webkit-appearance are not of interest to other UAs (see
https://github.com/whatwg/compat/issues/6), and the behavior is
different on different elements (e.g. <input>).

Since this has caused issues across different platforms on Chrome (see
bug), remove it.

Bug: 810162
Change-Id: I9d469cb624569f453978f3c56cc180eb07435b5b